Create `storage` directory for saving persistent data
This is a directory to save persistent data across `docker run`. All
files inside this directory are ignored by git. They shouldn't be
committed to the same git tree together with `server-a001` project.
The first use case of this directory is for the MySQL data. The
container will mount particular directories inside this `storage`
directory.

Side note: For data backup, the data inside the `storage` directory
should be backed up independently with this tree.

[ammarfaizi2: Reword the commit message]

Link: https://lore.gnuweeb.org/gwml/20220430170838.19571-4-ammarfaizi2@gnuweeb.org

Reviewed-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Reviewed-by: Hazmi Alfarizqi <hzmi@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>