start.sh: Fix readlink error when $MYSQL_DATA_DIR doesn't exist
There is a possibility that the command

    readlink -e "${MYSQL_DATA_DIR}"

gets executed when the directory ${MYSQL_DATA_DIR} doesn't exist.
When that happens, readlink command will error and we can't get the
absolute path of the MySQL data directory.

Fix this by executing mkdir -pv "${MYSQL_DATA_DIR}"; before the
readlink to ensure the directory exists.

Link: https://t.me/GNUWeeb/601793
Signed-off-by: Arthur Lapz <rlapz@gnuweeb.org>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>